+++ src/spawn.rs
@@ -173,12 +173,17 @@ pub struct Spawner {
 
     /// Time between bursts of the particle system, in seconds.
     ///
-    /// If this is infinity, there's only one burst.
     /// If this is [`spawn_duration`] or less, the system spawns a steady stream
     /// of particles.
     ///
     /// [`spawn_duration`]: Spawner::spawn_duration
     period: CpuValue<f32>,
+
+    /// Number of cycles the spawner is active before completing.
+    ///
+    /// Each cycle lasts for `period`. A value of `0` means "infinite", that is
+    /// the spanwe emits particle forever as long as it's active.
+    cycle_count: u32,
 
     /// Whether the spawner is active at startup.
     ///
@@ -186,23 +191,16 @@ pub struct Spawner {
     ///
     /// [`EffectSpawner::active`]: crate::EffectSpawner::active
     starts_active: bool,
-
-    /// Whether the burst of a once-style spawner triggers immediately when the
-    /// spawner becomes active.
-    ///
-    /// If `false`, the spawner doesn't do anything until
-    /// [`EffectSpawner::reset()`] is called.
-    starts_immediately: bool,
 }
 
 impl Default for Spawner {
     fn default() -> Self {
-        Self::once(1.0f32.into(), true)
+        Self::once(1.0f32.into())
     }
 }
 
 impl Spawner {
-    /// Create a spawner with a given count, time, and period.
+    /// Create a spawner.
     ///
     /// This is the _raw_ constructor. In general you should prefer using one of
     /// the utility constructors [`once()`], [`burst()`], or [`rate()`],
@@ -218,7 +216,9 @@ impl Spawner {
     ///   then the particles spawn all at once exactly at the same instant.
     /// - `period` is the amount of time between bursts of particles. If this is
     ///   <= `spawn_duration`, then the spawner spawns a steady stream of
-    ///   particles. If this is infinity, then there is a single burst.
+    ///   particles. This is ignored if `cycle_count == 1`.
+    /// - `cycle_count` is the number of times this pattern occurs. Set this to
+    ///   `0` to repeat forever while the spawner is active.
     ///
     /// ```txt
     ///  <----------- period ----------->
@@ -234,64 +234,74 @@ impl Spawner {
     /// single-frame burst, simply set the `spawn_duration` to zero; this is
     /// what [`once()`] does.
     ///
-    /// The `period` can be (positive) infinity; in that case, the spawner only
-    /// spawns a single time. This is equivalent to using [`once()`].
-    ///
     /// # Panics
     ///
     /// Panics if `period` can produce a negative number (the sample range lower
-    /// bound is negative), or can only produce 0 (the sample range upper bound
-    /// is not strictly positive).
+    /// bound is negative), unless the cycle count is exactly 1, in which case
+    /// the period is irrelevant.
+    ///
+    /// Panics if `period` can only produce 0 (the sample range upper bound
+    /// is not strictly positive), unless the cycle count is exactly 1, in which
+    /// case the period is irrelevant.
+    ///
+    /// Panics any value is infinite.
     ///
     /// # Example
     ///
     /// ```
     /// # use bevy_hanabi::Spawner;
     /// // Spawn 32 particles over 3 seconds, then pause for 7 seconds (10 - 3),
-    /// // and repeat.
-    /// let spawner = Spawner::new(32.0.into(), 3.0.into(), 10.0.into());
+    /// // doing that 5 times in total.
+    /// let spawner = Spawner::new(32.0.into(), 3.0.into(), 10.0.into(), 5);
     /// ```
     ///
     /// [`once()`]: crate::Spawner::once
     /// [`burst()`]: crate::Spawner::burst
     /// [`rate()`]: crate::Spawner::rate
-    pub fn new(count: CpuValue<f32>, spawn_duration: CpuValue<f32>, period: CpuValue<f32>) -> Self {
+    pub fn new(
+        count: CpuValue<f32>,
+        spawn_duration: CpuValue<f32>,
+        period: CpuValue<f32>,
+        cycle_count: u32,
+    ) -> Self {
         assert!(
-            period.range()[0] >= 0.,
+            cycle_count == 1 || period.range()[0] >= 0.,
             "`period` must not generate negative numbers (period.min was {}, expected >= 0).",
             period.range()[0]
         );
         assert!(
-            period.range()[1] > 0.,
+            cycle_count == 1 || period.range()[1] > 0.,
             "`period` must be able to generate a positive number (period.max was {}, expected > 0).",
             period.range()[1]
+        );
+        assert!(
+            period.range()[0].is_finite() && period.range()[1].is_finite(),
+            "`period` {:?} has an infinite bound. If upgrading from a previous version, use `cycle_count = 1` instead for a single-cycle burst.",
+            period
         );
 
         Self {
             count,
             spawn_duration,
             period,
+            cycle_count,
             starts_active: true,
-            starts_immediately: true,
         }
     }
 
     /// Create a spawner that spawns a burst of particles once.
     ///
     /// The burst of particles is spawned all at once in the same frame. After
-    /// that, the spawner idles, waiting to be manually reset via
-    /// [`EffectSpawner::reset()`].
+    /// that, the spawner deactivates, waiting to be manually reset via
+    /// [`EffectSpawner::reset()`] and re-activated with
+    /// [`EffectSpawner::set_active(true)`].
     ///
-    /// If `spawn_immediately` is `false`, this waits until
-    /// [`EffectSpawner::reset()`] before spawning a burst of particles.
-    ///
-    /// When `spawn_immediately == true`, this spawns a burst immediately on
-    /// activation. In that case, this is a convenience for:
+    /// This is a convenience for:
     ///
     /// ```
     /// # use bevy_hanabi::{Spawner, CpuValue};
     /// # let count = CpuValue::Single(1.);
-    /// Spawner::new(count, 0.0.into(), f32::INFINITY.into());
+    /// Spawner::new(count, 0.0.into(), 0.0.into(), 1);
     /// ```
     ///
     /// # Example
@@ -299,23 +309,26 @@ impl Spawner {
     /// ```
     /// # use bevy_hanabi::Spawner;
     /// // Spawn 32 particles in a burst once immediately on creation.
-    /// let spawner = Spawner::once(32.0.into(), true);
+    /// let spawner = Spawner::once(32.0.into());
     /// ```
     ///
-    /// [`reset()`]: crate::Spawner::reset
-    pub fn once(count: CpuValue<f32>, spawn_immediately: bool) -> Self {
-        let mut spawner = Self::new(count, 0.0.into(), f32::INFINITY.into());
-        spawner.starts_immediately = spawn_immediately;
-        spawner
+    /// [`EffectSpawner::set_active(true)`]: crate::EffectSpawner::set_active
+    pub fn once(count: CpuValue<f32>) -> Self {
+        Self::new(count, 0.0.into(), 0.0.into(), 1)
     }
 
     /// Get whether this spawner emits a single burst.
+    ///
+    /// This is true if the cycle count is exactly equal to 1.
     pub fn is_once(&self) -> bool {
-        if let CpuValue::Single(f) = self.period {
-            f.is_infinite()
-        } else {
-            false
-        }
+        self.cycle_count == 1
+    }
+
+    /// Get whether this spawner emits particles forever while active.
+    ///
+    /// This is true if the cycle count is exactly equal to 0.
+    pub fn is_forever(&self) -> bool {
+        self.cycle_count == 0
     }
 
     /// Create a spawner that spawns particles at `rate`, accumulated each
@@ -326,7 +339,7 @@ impl Spawner {
     /// ```
     /// # use bevy_hanabi::{Spawner, CpuValue};
     /// # let rate = CpuValue::Single(1.);
-    /// Spawner::new(rate, 1.0.into(), 1.0.into());
+    /// Spawner::new(rate, 1.0.into(), 1.0.into(), 0);
     /// ```
     ///
     /// # Example
@@ -337,7 +350,7 @@ impl Spawner {
     /// let spawner = Spawner::rate(10.0.into());
     /// ```
     pub fn rate(rate: CpuValue<f32>) -> Self {
-        Self::new(rate, 1.0.into(), 1.0.into())
+        Self::new(rate, 1.0.into(), 1.0.into(), 0)
     }
 
     /// Create a spawner that spawns `count` particles, waits `period` seconds,
@@ -349,7 +362,7 @@ impl Spawner {
     /// # use bevy_hanabi::{Spawner, CpuValue};
     /// # let count = CpuValue::Single(1.);
     /// # let period = CpuValue::Single(1.);
-    /// Spawner::new(count, 0.0.into(), period);
+    /// Spawner::new(count, 0.0.into(), period, 0);
     /// ```
     ///
     /// # Example
@@ -360,7 +373,7 @@ impl Spawner {
     /// let spawner = Spawner::burst(5.0.into(), 3.0.into());
     /// ```
     pub fn burst(count: CpuValue<f32>, period: CpuValue<f32>) -> Self {
-        Self::new(count, 0.0.into(), period)
+        Self::new(count, 0.0.into(), period, 0)
     }
 
     /// Set the number of particles that are spawned each cycle.
@@ -397,33 +410,85 @@ impl Spawner {
 
     /// Set the duration of a spawn cycle, in seconds.
     ///
-    /// A spawn cycles includes the [`spawn_duration()`] value, and any extra
+    /// A spawn cycle includes the [`spawn_duration()`] value, and any extra
     /// wait time (if larger than spawn time).
+    ///
+    /// # Panics
+    ///
+    /// Panics if the period is infinite.
     ///
     /// [`spawn_duration()`]: Self::spawn_duration
     pub fn with_period(mut self, period: CpuValue<f32>) -> Self {
+        assert!(
+            period.range()[0].is_finite() && period.range()[1].is_finite(),
+            "`period` {:?} has an infinite bound. If upgrading from a previous version, use `cycle_count = 1` instead for a single-cycle burst.",
+            period
+        );
         self.period = period;
         self
     }
 
     /// Set the duration of the spawn cycle, in seconds.
     ///
-    /// A spawn cycles includes the [`spawn_duration()`] value, and any extra
+    /// A spawn cycle includes the [`spawn_duration()`] value, and any extra
     /// wait time (if larger than spawn time).
+    ///
+    /// # Panics
+    ///
+    /// Panics if the period is infinite.
     ///
     /// [`spawn_duration()`]: Self::spawn_duration
     pub fn set_period(&mut self, period: CpuValue<f32>) {
+        assert!(
+            period.range()[0].is_finite() && period.range()[1].is_finite(),
+            "`period` {:?} has an infinite bound. If upgrading from a previous version, use `cycle_count = 1` instead for a single-cycle burst.",
+            period
+        );
         self.period = period;
     }
 
     /// Get the duration of the spawn cycle, in seconds.
     ///
-    /// A spawn cycles includes the [`spawn_duration()`] value, and any extra
+    /// A spawn cycle includes the [`spawn_duration()`] value, and any extra
     /// wait time (if larger than spawn time).
     ///
     /// [`spawn_duration()`]: Self::spawn_duration
     pub fn period(&self) -> CpuValue<f32> {
         self.period
+    }
+
+    /// Set the number of cycles to spawn for.
+    ///
+    /// A spawn cycle includes the [`spawn_duration()`] value, and any extra
+    /// wait time (if larger than spawn time). It lasts for [`period()`].
+    ///
+    /// [`spawn_duration()`]: Self::spawn_duration
+    /// [`period()`]: Self::period
+    pub fn with_cycle_count(mut self, cycle_count: u32) -> Self {
+        self.cycle_count = cycle_count;
+        self
+    }
+
+    /// Set the number of cycles to spawn for.
+    ///
+    /// A spawn cycle includes the [`spawn_duration()`] value, and any extra
+    /// wait time (if larger than spawn time). It lasts for [`period()`].
+    ///
+    /// [`spawn_duration()`]: Self::spawn_duration
+    /// [`period()`]: Self::period
+    pub fn set_cycle_count(&mut self, cycle_count: u32) {
+        self.cycle_count = cycle_count;
+    }
+
+    /// Get the number of cycles to spawn for.
+    ///
+    /// A spawn cycle includes the [`spawn_duration()`] value, and any extra
+    /// wait time (if larger than spawn time). It lasts for [`period()`].
+    ///
+    /// [`spawn_duration()`]: Self::spawn_duration
+    /// [`period()`]: Self::period
+    pub fn cycle_count(&self) -> u32 {
+        self.cycle_count
     }
 
     /// Sets whether the spawner starts active when the effect is instantiated.
@@ -461,20 +526,26 @@ impl Spawner {
 pub struct EffectSpawner {
     /// The spawner configuration extracted from the [`EffectAsset`], or
     /// directly overriden by the user.
-    spawner: Spawner,
+    pub spawner: Spawner,
 
-    /// Accumulated time since last spawn, in seconds.
-    time: f32,
+    /// Accumulated time for the current (partial) cycle, in seconds.
+    cycle_time: f32,
+
+    /// Number of cycles already completed.
+    completed_cycle_count: u32,
 
     /// Sampled value of `spawn_duration` until `period` is reached. This is the
     /// duration of the "active" period during which we spawn particles, as
     /// opposed to the "wait" period during which we do nothing until the next
     /// spawn cycle.
-    spawn_duration: f32,
+    sampled_spawn_duration: f32,
 
     /// Sampled value of the time period, in seconds, until the next spawn
     /// cycle.
-    period: f32,
+    sampled_period: f32,
+
+    /// Sampled value of the number of particles to spawn per `spawn_duration`.
+    sampled_count: f32,
 
     /// Number of particles to spawn this frame.
     ///
@@ -485,7 +556,7 @@ pub struct EffectSpawner {
     /// You can manually assign this value to override the one calculated by
     /// [`tick()`]. Note in this case that you need to override the value after
     /// the automated one was calculated, by ordering your system
-    /// after [`tick_initializers()`] or [`EffectSystems::TickSpawners`].
+    /// after [`tick_spawners()`] or [`EffectSystems::TickSpawners`].
     ///
     /// [`tick()`]: crate::EffectSpawner::tick
     /// [`EffectSystems::TickSpawners`]: crate::EffectSystems::TickSpawners
@@ -507,13 +578,11 @@ impl EffectSpawner {
     pub fn new(spawner: &Spawner) -> Self {
         Self {
             spawner: *spawner,
-            time: if spawner.is_once() && !spawner.starts_immediately {
-                1. // anything > 0
-            } else {
-                0.
-            },
-            spawn_duration: 0.,
-            period: 0.,
+            cycle_time: 0.,
+            completed_cycle_count: 0,
+            sampled_spawn_duration: 0.,
+            sampled_period: 0.,
+            sampled_count: 0.,
             spawn_count: 0,
             spawn_remainder: 0.,
             active: spawner.starts_active(),
@@ -531,6 +600,7 @@ impl EffectSpawner {
     /// Set whether the spawner is active.
     ///
     /// Inactive spawners do not tick, and therefore do not spawn any particle.
+    #[inline]
     pub fn set_active(&mut self, active: bool) {
         self.active = active;
     }
@@ -538,17 +608,75 @@ impl EffectSpawner {
     /// Get whether the spawner is active.
     ///
     /// Inactive spawners do not tick, and therefore do not spawn any particle.
+    #[inline]
     pub fn is_active(&self) -> bool {
         self.active
     }
 
-    /// Get the spawner configuration in use.
+    /// Get the time relative to the beginning of the current cycle.
+    #[inline]
+    pub fn cycle_time(&self) -> f32 {
+        self.cycle_time
+    }
+
+    /// Get the spawn duration for the current cycle.
     ///
-    /// The effective [`Spawner`] used is either the override specified in the
-    /// associated [`ParticleEffect`] instance, or the fallback one specified in
-    /// underlying [`EffectAsset`].
-    pub fn spawner(&self) -> &Spawner {
-        &self.spawner
+    /// This value can change every cycle if [`Spawner::spawn_duration`] is a
+    /// randomly distributed value.
+    #[inline]
+    pub fn cycle_spawn_duration(&self) -> f32 {
+        self.sampled_spawn_duration
+    }
+
+    /// Get the period of the current cycle.
+    ///
+    /// This value can change every cycle if [`Spawner::period`] is a randomly
+    /// distributed value. If the effect spawns only once, and therefore its
+    /// cycle period is ignored, this returns `0`.
+    #[inline]
+    pub fn cycle_period(&self) -> f32 {
+        if self.spawner.is_once() {
+            0.
+        } else {
+            self.sampled_period
+        }
+    }
+
+    /// Get the progress ratio in 0..1 of the current cycle.
+    ///
+    /// This is the ratio of the [`cycle_time()`] over [`cycle_period()`]. If
+    /// the effect spawns only once, and therefore its cycle period is
+    /// ignored, this returns `0`.
+    ///
+    /// [`cycle_time()`]: Self::cycle_time
+    /// [`cycle_period()`]: Self::cycle_period
+    #[inline]
+    pub fn cycle_ratio(&self) -> f32 {
+        if self.spawner.is_once() {
+            0.
+        } else {
+            self.cycle_time / self.sampled_period
+        }
+    }
+
+    /// Get the number of particles to spawn during the current cycle
+    ///
+    /// This value can change every cycle if [`Spawner::count`] is a randomly
+    /// distributed value.
+    #[inline]
+    pub fn cycle_spawn_count(&self) -> f32 {
+        self.sampled_count
+    }
+
+    /// Get the number of completed cycles since last [`reset()`].
+    ///
+    /// The value loops back if the pattern repeats forever
+    /// ([`Spawner::is_forever()`] is `true`).
+    ///
+    /// [`reset()`]: Self::reset
+    #[inline]
+    pub fn completed_cycle_count(&self) -> u32 {
+        self.completed_cycle_count
     }
 
     /// Reset the spawner state.
@@ -561,8 +689,11 @@ impl EffectSpawner {
     ///
     /// [`Spawner::once`]: crate::Spawner::once
     pub fn reset(&mut self) {
-        self.time = 0.;
-        self.period = 0.;
+        self.cycle_time = 0.;
+        self.completed_cycle_count = 0;
+        self.sampled_spawn_duration = 0.;
+        self.sampled_period = 0.;
+        self.sampled_count = 0.;
         self.spawn_count = 0;
         self.spawn_remainder = 0.;
     }
@@ -582,59 +713,90 @@ impl EffectSpawner {
     /// The integral number of particles to spawn this frame. Any fractional
     /// remainder is saved for the next call.
     pub fn tick(&mut self, mut dt: f32, rng: &mut Pcg32) -> u32 {
+        // Handle max cycle count here, which might have been reached last frame, but we
+        // couldn't deactivate before the end of the frame.
+        if !self.spawner.is_forever() && (self.completed_cycle_count >= self.spawner.cycle_count())
+        {
+            self.active = false;
+        }
+
         if !self.active {
             self.spawn_count = 0;
             return 0;
         }
 
-        // The limit can be reached multiple times, so use a loop
+        // Use a loop in case the timestep dt spans multiple cycles
         loop {
-            if self.period == 0.0 {
-                self.resample(rng);
-                continue;
+            // Check if this is a new cycle which needs resampling
+            if self.sampled_period == 0.0 {
+                if self.spawner.is_once() {
+                    self.sampled_spawn_duration = self.spawner.spawn_duration.sample(rng);
+                    // Period is unchecked, should be ignored (could sample to <= 0). Use the spawn
+                    // duration, but ensure we have something > 0 as a marker that we've resampled.
+                    self.sampled_period = self.sampled_spawn_duration.max(1e-12);
+                } else {
+                    self.sampled_period = self.spawner.period.sample(rng);
+                    assert!(self.sampled_period > 0.);
+                    self.sampled_spawn_duration = self
+                        .spawner
+                        .spawn_duration
+                        .sample(rng)
+                        .clamp(0., self.sampled_period);
+                }
+                self.sampled_spawn_duration = self.spawner.spawn_duration.sample(rng);
+                self.sampled_count = self.spawner.count.sample(rng).max(0.);
             }
 
-            let new_time = self.time + dt;
-            if self.time <= self.spawn_duration {
+            let new_time = self.cycle_time + dt;
+
+            // If inside the spawn period, accumulate some particle spawn count
+            if self.cycle_time <= self.sampled_spawn_duration {
                 // If the spawn time is very small, close to zero, spawn all particles
                 // immediately in one burst over a single frame.
-                self.spawn_remainder += if self.spawn_duration < 1e-5f32.max(dt / 100.0) {
-                    self.spawner.count.sample(rng)
+                self.spawn_remainder += if self.sampled_spawn_duration < 1e-5f32.max(dt / 100.0) {
+                    self.sampled_count
                 } else {
                     // Spawn an amount of particles equal to the fraction of time the current frame
                     // spans compared to the total burst duration.
-                    self.spawner.count.sample(rng) * (new_time.min(self.spawn_duration) - self.time)
-                        / self.spawn_duration
+                    let ratio = ((new_time.min(self.sampled_spawn_duration) - self.cycle_time)
+                        / self.sampled_spawn_duration)
+                        .clamp(0., 1.);
+                    self.sampled_count * ratio
                 };
             }
 
-            let old_time = self.time;
-            self.time = new_time;
+            // Increment current time
+            self.cycle_time = new_time;
 
-            if self.time >= self.period {
-                dt -= self.period - old_time;
-                self.time = 0.0; // dt will be added on in the next iteration
-                self.resample(rng);
+            // Check for cycle completion
+            if self.cycle_time >= self.sampled_period {
+                dt = self.cycle_time - self.sampled_period;
+                self.cycle_time = 0.0;
+                self.completed_cycle_count += 1;
+
+                // Mark as "need resampling"
+                self.sampled_period = 0.0;
+
+                // If this was the last cycle, we're done
+                if !self.spawner.is_forever()
+                    && (self.completed_cycle_count >= self.spawner.cycle_count())
+                {
+                    // Don't deactivate quite yet, otherwise we'll miss the spawns for this frame
+                    break;
+                }
             } else {
+                // We're done for this frame
                 break;
             }
         }
 
+        // Extract integral number of particles to spawn this frame, keep remainder for
+        // next one
         let count = self.spawn_remainder.floor();
         self.spawn_remainder -= count;
         self.spawn_count = count as u32;
 
         self.spawn_count
-    }
-
-    /// Resamples the spawn time and period.
-    fn resample(&mut self, rng: &mut Pcg32) {
-        self.period = self.spawner.period.sample(rng);
-        self.spawn_duration = self
-            .spawner
-            .spawn_duration
-            .sample(rng)
-            .clamp(0.0, self.period);
     }
 }
 
@@ -668,7 +830,7 @@ pub fn tick_spawners(
     mut query: Query<(
         Entity,
         &ParticleEffect,
-        Option<&InheritedVisibility>,
+        &InheritedVisibility,
         Option<&mut EffectSpawner>,
     )>,
 ) {
@@ -676,7 +838,7 @@ pub fn tick_spawners(
 
     let dt = time.delta_secs();
 
-    for (entity, effect, maybe_inherited_visibility, maybe_spawner) in query.iter_mut() {
+    for (entity, effect, inherited_visibility, maybe_spawner) in query.iter_mut() {
         let Some(asset) = effects.get(&effect.handle) else {
             trace!(
                 "Effect asset with handle {:?} is not available; skipped initializers tick.",
@@ -686,9 +848,7 @@ pub fn tick_spawners(
         };
 
         if asset.simulation_condition == SimulationCondition::WhenVisible
-            && !maybe_inherited_visibility
-                .map(|iv| iv.get())
-                .unwrap_or(true)
+            && !inherited_visibility.get()
         {
             trace!(
                 "Effect asset with handle {:?} is not visible, and simulates only WhenVisible; skipped initializers tick.",
@@ -751,35 +911,66 @@ mod test {
     #[test]
     fn test_new() {
         let rng = &mut new_rng();
-        // 3 particles over 3 seconds, pause 7 seconds (total 10 seconds period).
-        let spawner = Spawner::new(3.0.into(), 3.0.into(), 10.0.into());
+        // 3 particles over 3 seconds, pause 7 seconds (total 10 seconds period). 2
+        // cycles.
+        let spawner = Spawner::new(3.0.into(), 3.0.into(), 10.0.into(), 2);
         let mut spawner = EffectSpawner::new(&spawner);
-        let count = spawner.tick(2.0, rng); // t = 2s
+        let count = spawner.tick(2., rng); // t = 2s
         assert_eq!(count, 2);
-        let count = spawner.tick(5.0, rng); // t = 7s
+        assert!(spawner.is_active());
+        assert_eq!(spawner.cycle_time(), 2.);
+        assert_eq!(spawner.cycle_spawn_duration(), 3.);
+        assert_eq!(spawner.cycle_period(), 10.);
+        assert_eq!(spawner.cycle_ratio(), 0.2); // 2s / 10s
+        assert_eq!(spawner.cycle_spawn_count(), 3.);
+        assert_eq!(spawner.completed_cycle_count(), 0);
+        let count = spawner.tick(5., rng); // t = 7s
         assert_eq!(count, 1);
-        let count = spawner.tick(8.0, rng); // t = 15s
+        assert!(spawner.is_active());
+        assert_eq!(spawner.cycle_time(), 7.);
+        assert_eq!(spawner.cycle_spawn_duration(), 3.);
+        assert_eq!(spawner.cycle_period(), 10.);
+        assert_eq!(spawner.cycle_ratio(), 0.7); // 7s / 10s
+        assert_eq!(spawner.cycle_spawn_count(), 3.);
+        assert_eq!(spawner.completed_cycle_count(), 0);
+        let count = spawner.tick(8., rng); // t = 15s
         assert_eq!(count, 3);
+        assert!(spawner.is_active());
+        assert_eq!(spawner.cycle_time(), 5.); // 15. mod 10.
+        assert_eq!(spawner.cycle_spawn_duration(), 3.);
+        assert_eq!(spawner.cycle_period(), 10.);
+        assert_eq!(spawner.cycle_ratio(), 0.5); // 5s / 10s
+        assert_eq!(spawner.cycle_spawn_count(), 3.);
+        assert_eq!(spawner.completed_cycle_count(), 1);
+        let count = spawner.tick(10., rng); // t = 25s
+        assert_eq!(count, 0);
+        assert!(spawner.is_active()); // still active for one frame
+        assert_eq!(spawner.completed_cycle_count(), 2);
+        let count = spawner.tick(0.1, rng); // t = 25.1s
+        assert_eq!(count, 0);
+        assert!(!spawner.is_active());
+        assert_eq!(spawner.completed_cycle_count(), 2);
     }
 
     #[test]
     #[should_panic]
     fn test_new_panic_negative_period() {
-        let _ = Spawner::new(3.0.into(), 1.0.into(), CpuValue::Uniform((-1., 1.)));
+        let _ = Spawner::new(3.0.into(), 1.0.into(), CpuValue::Uniform((-1., 1.)), 0);
     }
 
     #[test]
     #[should_panic]
     fn test_new_panic_zero_period() {
-        let _ = Spawner::new(3.0.into(), 1.0.into(), CpuValue::Uniform((0., 0.)));
+        let _ = Spawner::new(3.0.into(), 1.0.into(), CpuValue::Uniform((0., 0.)), 0);
     }
 
     #[test]
     fn test_once() {
         let rng = &mut new_rng();
-        let spawner = Spawner::once(5.0.into(), true);
+        let spawner = Spawner::once(5.0.into());
         assert!(spawner.is_once());
         let mut spawner = EffectSpawner::new(&spawner);
+        assert!(spawner.is_active());
         let count = spawner.tick(0.001, rng);
         assert_eq!(count, 5);
         let count = spawner.tick(100.0, rng);
@@ -789,8 +980,9 @@ mod test {
     #[test]
     fn test_once_reset() {
         let rng = &mut new_rng();
-        let spawner = Spawner::once(5.0.into(), true);
+        let spawner = Spawner::once(5.0.into());
         assert!(spawner.is_once());
+        assert!(spawner.starts_active());
         let mut spawner = EffectSpawner::new(&spawner);
         spawner.tick(1.0, rng);
         spawner.reset();
@@ -799,16 +991,41 @@ mod test {
     }
 
     #[test]
-    fn test_once_not_immediate() {
+    fn test_once_start_inactive() {
         let rng = &mut new_rng();
-        let spawner = Spawner::once(5.0.into(), false);
+
+        let spawner = Spawner::once(5.0.into()).with_starts_active(false);
         assert!(spawner.is_once());
+        assert!(!spawner.starts_active());
         let mut spawner = EffectSpawner::new(&spawner);
+
+        // Inactive; no-op
         let count = spawner.tick(1.0, rng);
         assert_eq!(count, 0);
-        spawner.reset();
+
+        spawner.set_active(true);
+
+        // Active; spawns
         let count = spawner.tick(1.0, rng);
         assert_eq!(count, 5);
+        assert!(spawner.is_active());
+
+        // Completed; deactivated itself
+        let count = spawner.tick(1.0, rng);
+        assert_eq!(count, 0);
+        assert!(!spawner.is_active());
+
+        // Reset internal state, but doesn't activate
+        spawner.reset();
+        let count = spawner.tick(1.0, rng);
+        assert_eq!(count, 0);
+        assert!(!spawner.is_active());
+
+        // Now activate again
+        spawner.set_active(true);
+        let count = spawner.tick(1.0, rng);
+        assert_eq!(count, 5);
+        assert!(spawner.is_active());
     }
 
     #[test]
@@ -816,6 +1033,7 @@ mod test {
         let rng = &mut new_rng();
         let spawner = Spawner::rate(5.0.into());
         assert!(!spawner.is_once());
+        assert!(spawner.is_forever());
         let mut spawner = EffectSpawner::new(&spawner);
         // Slightly over 1.0 to avoid edge case
         let count = spawner.tick(1.01, rng);
@@ -857,6 +1075,7 @@ mod test {
         let rng = &mut new_rng();
         let spawner = Spawner::burst(5.0.into(), 2.0.into());
         assert!(!spawner.is_once());
+        assert!(spawner.is_forever());
         let mut spawner = EffectSpawner::new(&spawner);
         let count = spawner.tick(1.0, rng);
         assert_eq!(count, 5);
@@ -941,7 +1160,7 @@ mod test {
 
     #[test]
     fn test_tick_spawners() {
-        let asset_spawner = Spawner::once(32.0.into(), true);
+        let asset_spawner = Spawner::once(32.0.into());
 
         for test_case in &[
             TestCase::new(None, asset_spawner),
@@ -989,7 +1208,7 @@ mod test {
             };
 
             // Tick once
-            let cur_time = {
+            let _cur_time = {
                 // Make sure to increment the current time so that the spawners spawn something.
                 // Note that `Time` has this weird behavior where the common quantities like
                 // `Time::delta_secs()` only update after the *second* update. So we tick the
@@ -1032,12 +1251,13 @@ mod test {
                     let actual_spawner = effect_spawner.spawner;
 
                     // Check the spawner ticked
-                    assert!(effect_spawner.active);
+                    assert!(effect_spawner.active); // will get deactivated next tick()
                     assert_eq!(effect_spawner.spawn_remainder, 0.);
-                    assert_eq!(effect_spawner.time, cur_time.as_secs_f32());
+                    assert_eq!(effect_spawner.cycle_time, 0.);
+                    assert_eq!(effect_spawner.completed_cycle_count, 1);
+                    assert_eq!(effect_spawner.spawn_count, 32);
 
                     assert_eq!(actual_spawner, test_case.asset_spawner);
-                    assert_eq!(effect_spawner.spawn_count, 32);
                 } else {
                     // If not visible, `tick_initializers()` skips the effect entirely so won't
                     // spawn an `EffectSpawner` for it
@@ -1059,12 +1279,13 @@ mod test {
                 let actual_spawner = effect_spawner.spawner;
 
                 // Check the spawner ticked
-                assert!(effect_spawner.active);
+                assert!(effect_spawner.active); // will get deactivated next tick()
                 assert_eq!(effect_spawner.spawn_remainder, 0.);
-                assert_eq!(effect_spawner.time, cur_time.as_secs_f32());
+                assert_eq!(effect_spawner.cycle_time, 0.);
+                assert_eq!(effect_spawner.completed_cycle_count, 1);
+                assert_eq!(effect_spawner.spawn_count, 32);
 
                 assert_eq!(actual_spawner, test_case.asset_spawner);
-                assert_eq!(effect_spawner.spawn_count, 32);
             }
         }
     }