+++ CHANGELOG.md
@@ -39,6 +39,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Previously the PRNG seed was implicitly set to a random value.
   To restore the former behavior, just set `prng_seed = rand::random::<u32>()`.
 - Added some new expressions: `acos`, `asin`, `atan`, `atan2`, `round`.
+- Added `Spawner::cycle_count`, which allows repeating a spawn pattern exactly N times (or forever if N=0).
+  Once the cycle count is reached, the spawner deactivates itself at the beginning of the next `tick()` call.
+- Added several utility functions on `EffectSpawner` to determine its current state.
 
 ### Changed
 
@@ -57,6 +60,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   nor any other built-in functionality of Hanabi.
 - `ParticleEffect` now requires (in the ECS sense) the `CompiledParticleEffect`, `Visibility`, and `Transform` components.
   This means those components are automatically added by Bevy each time a `ParticleEffect` is spawned. (#418)
+- The `EffectSpawner::spawner` field is now public.
+- `Spawner::once()` will now deactivate the spawner after the one (and only) burst.
+  This makes the behavior consistent with the new loop count,
+  and also enforces the semantic that an active spawner always has more particles to spawn.
 
 ### Fixed
 
@@ -81,6 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Use the `Attribute::RIBBON_ID` instead to assign a per-particle ribbon ID.
 - Removed `ParticleEffectBundle`. Use `ParticleEffect` directly instead.
 - Removed `ParticleEffect::z_layer_2d`. Use the Z coordinate of the effect's `Tranform` to order effects. (#423)
+- Removed `EffectSpawner::spawner()` as the `spawner` field is now public.
+- Removed `Spawner::starts_immediately` which was not working and was duplicating semantic with `active`. (#420)
 
 ## [0.14.0] 2024-12-09